+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - ACL: Fix issues with corrputed imports
 - Animals: Delay vote embed creation to combat Discord-side issues with content not being available at event creation
 - Base: Prevent users from pinning system messages
+- Comment: User comment module
 - Karma: Limit "karma emoji" command to guild contexts
 - Librarian: Test "ipaddress" query with regex
 - Random: Include explainxkcd.com link in "xkcd" command